fix: remove old daytime utils (hacky, incorrect) [PT-185589271]
