+++ star-navigation/src/utils/sim-utils.ts
@@ -53,6 +53,13 @@ export const getTimezone = (month: number, day: number) => {
   return PST;
 };
 
+export const dateToFractionalHoursInRightTimezone = (date: Date) => {
+  const sunriseInTimezone = new Date(date.toLocaleString("en-US", {
+    timeZone: "America/Anchorage"
+  }));
+  return sunriseInTimezone.getHours() + sunriseInTimezone.getMinutes() / 60;
+};
+
 export const getDateTimeString = ({ month, day, timeOfDay } : { month: number, day: number, timeOfDay: number}) =>
   `${SIMULATION_YEAR}-${formatTimeNumber(month)}-${formatTimeNumber(day)}T${fractionalHourToTimeString(timeOfDay)}${getTimezone(month, day)}`;
